Prepare import of screen 480x320 (#732)

+++ src/gui/gui.h
@@ -188,7 +188,7 @@ typedef struct guiButton {
 typedef struct guiScrollable {
     struct guiHeader header;
     u8 item_count;
-    u8 row_height;
+    u16 row_height;
     u8 selectable_rows;
     u8 cur_row;
     u8 visible_rows;
@@ -362,7 +362,7 @@ void GUI_ButtonEnable(guiObject_t *obj, u8 enable);
 unsigned GUI_IsButtonEnabled(guiObject_t *obj);
 unsigned GUI_IsButtonLongPress(guiObject_t *obj);
 
-guiObject_t *GUI_CreateScrollable(guiScrollable_t *scrollable, u16 x, u16 y, u16 width, u16 height, u8 row_height, u8 item_count,
+guiObject_t *GUI_CreateScrollable(guiScrollable_t *scrollable, u16 x, u16 y, u16 width, u16 height, u16 row_height, u8 item_count,
      int (*row_cb)(int absrow, int relrow, int x, void *data),
      guiObject_t * (*getobj_cb)(int relrow, int col, void *data),
      int (*size_cb)(int absrow, void *data),
